backend/services/recruitment_agent/requirements.txt
python-jose[cryptography]==3.3.0

# Core Web
fastapi==0.111.0
uvicorn==0.28.1
# Database
sqlalchemy==2.0.29
pymysql==1.1.0

# Validation
pydantic==2.7.1
python-multipart==0.0.9

# PDF Parsing
PyMuPDF==1.24.0

# AI / LLM
langchain==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.14
langgraph==0.2.70
openai==1.55.3

# Utils
requests==2.31.0
python-dotenv==1.0.1